django_registry/requirements.txt
Django>=5.1,<6
djangorestframework>=3.15,<4
psycopg2-binary>=2.9,<3
